requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
mysqlclient>=2.1
gunicorn>=20.1
whitenoise>=6.4
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
pillow>=10.0
